Move Header component to App.js and move logic for calling fetchUser action to App.js
